refactor: further reduced size of the main redis service. extracted most of the main, important functionality in modules of their own which turns the main redis service into much more of a facade than it used to be. now it has almost no reason to change except for wiring and it acts as an API, receiving requests and routing them to the appropriate submodule
